requirements.txt
streamlit
pandas
numpy
scikit-learn
textblob
matplotlib
plotly
supabase
python-dotenv
